Refactor platform and URL utilities to separate classes
Moved platform-specific path resolution and library existence checks to PlatformPathResolver, and YouTube URL utility methods to UrlUtilities. Updated all usages across the codebase to use these new utility classes, improving modularity and maintainability.
